+++ utils.js
@@ -81,13 +81,11 @@ utils.parseOutCode = function( fileContents, regex ) {
   // Removes the first item in the array since it will always be empty.
   code.shift();
   // Clean each item in the array.
-  code.forEach(function( item, index ){
-    code[ index ] = code[ index ].trim();
-  });
+  code = code.map((item) => item.replace(regex.ignore, '').trim());
   return code;
 };
 
-utils.removeDoxrayCommentTokens = function( item, regex ) {
+utils.removeDoxrayCommentTokens = function ( item, regex ) {
   // Remove the opening and closing comments.
   return item.replace( regex.opening, '' ).replace( regex.closing, '' );
 };